backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
pymysql==1.1.0
flask-migrate==4.0.4
requests==2.31.0
flask_session==0.5.0
cryptography==43.0.0
pypinyin==0.50.0